Calls a Splunk REST endpoint to reload the views when done fetching data for all the repositories.
The call to the REST endpoint is needed because
the view won't show up automatically in Splunk 
when it has been edited outside of Splunk.

In addition, to call this endpoint, we need the
Splunk username and password. So I also created a
python script that uses splunkgit_settings.py to
print the configured user and password.