Fixed bug where --skip in git log didn't do what we thought.
We used the --skip flag to git log to avoid
duplicate data. The problem is that it skipped
the latest commits, which resulted in that we didn't
get any new commits, until we restarted and cleaned
Splunk.

Now we use --since=<date> and <since>.. instead,
which will avoid duplicates and get the commits
after the "since commit", which is the commit with
the most recent time, indexed in Splunk.