Changed the Splunk search for how we skip commits
Now does a Splunk search to see if there are any
git_file_changes to a specific repo, instead of
distinct counting all the commit_hashes of the
whole repository. The Splunk search is for indexing
new commits with the git log command.

This change and the --skip flag change makes the
script many magnitudes faster on large repositories.
Which also helps when dealing with real-time updates.